+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+ - Added `SamplesPassedQuery` and `TimeElapsedQuery` types. They can be passed to the draw parameters.
  - Buffers are no longer mapped with `GL_MAP_COHERENT_BIT`. Flushing is handled by glium.
  - Changed `Surface::clear` to take an additional optional `Rect` parameter that specifies the rectangle to clear.
  - Fixed the `program!` macro not usable with version numbers >= 256.